style: rename the `classnames.ts`
