+++ utils/src/index.ts
@@ -1,3 +1,3 @@
 export * from './types'
 
-export { createCls } from './classnames'
+export { createCls } from './class-names'
